requirements.txt
Flask
pandas
numpy
plotly
plotly-express
python-dotenv
redis
psycopg2
openpyxl
